test: fix imports for CI
