+++ e2e/desktop/tests/specs/ui.swap.spec.ts
@@ -7,7 +7,7 @@ import { Provider } from "@ledgerhq/live-common/e2e/enum/Provider";
 import { addTmsLink } from "tests/utils/allureUtils";
 import { getDescription } from "tests/utils/customJsonReporter";
 import { setupEnv, performSwapUntilQuoteSelectionStep } from "tests/utils/swapUtils";
-import { liveDataWithAddressCommand } from "@ledgerhq/live-common/e2e";
+import { liveDataWithAddressCommand } from "@ledgerhq/live-common/e2e/cliCommandsUtils";
 
 const exchangeAppInfo: AppInfos = AppInfos.EXCHANGE;
 const swapUiTags = [
