+++ libs/ledger-live-common/.unimportedrc.json
@@ -60,7 +60,9 @@
     "src/domain/getTotalStakeableAssets.ts",
     "src/e2e/data/assetsDrawer.ts",
     "src/e2e/data/regexes.ts",
+    "src/e2e/cliCommandsUtils.ts",
     "src/e2e/index.ts",
+    "src/e2e/runCli.ts",
     "src/e2e/speculos.ts",
     "src/e2e/speculosAppVersion.ts",
     "src/e2e/swap.ts",
